Add variables and function, remove text from div.
Added "chancesLeft" and "rightGuesses" variables, to be used for score
keeping and to trigger the end of the game. Added function endGame, with
an overall outline for what the function should do. Removed score texts
from inside div (in html), instead, the divs have appropriate id's, and
the score should be updated in each round (through the script).