Add HTML elements for every stage of the game.
Added HTML elements for the stages of the game, every stage is in a div
and has a class, the idea is to use classList.toggle to hide and show
these elements according to the state of the game. In relation to the
dynamic of the game, decided to change the narrative (as seen in first
<p> element), won't have an input to "make a wish" anymore, if you win,
the game says the genie will fix your bug.